requirements.txt
Flask
Pillow
pikepdf
python-docx
openpyxl
mutagen
